+++ src/librustc/lib.rs
@@ -39,7 +39,6 @@
 #![feature(unsafe_destructor)]
 #![feature(staged_api)]
 #![feature(std_misc)]
-#![feature(unicode)]
 #![feature(os)]
 #![cfg_attr(test, feature(test))]
