+++ src/librustdoc/lib.rs
@@ -42,6 +42,7 @@ extern crate rustc;
 extern crate rustc_trans;
 extern crate rustc_driver;
 extern crate rustc_resolve;
+extern crate rustc_lint;
 extern crate serialize;
 extern crate syntax;
 extern crate "test" as testing;
